Move utility method getExpectedRetentionTimeInMs from TopicManager to StoreUtils class
