+++ internal/venice-common/src/main/java/com/linkedin/venice/meta/ZKStore.java
@@ -5,11 +5,11 @@ import com.linkedin.venice.common.VeniceSystemStoreType;
 import com.linkedin.venice.compression.CompressionStrategy;
 import com.linkedin.venice.exceptions.StoreDisabledException;
 import com.linkedin.venice.exceptions.VeniceException;
-import com.linkedin.venice.pubsub.manager.TopicManager;
 import com.linkedin.venice.systemstore.schemas.StoreProperties;
 import com.linkedin.venice.systemstore.schemas.StoreVersion;
 import com.linkedin.venice.utils.AvroCompatibilityUtils;
 import com.linkedin.venice.utils.AvroRecordUtils;
+import com.linkedin.venice.utils.StoreUtils;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -741,7 +741,7 @@ public class ZKStore extends AbstractStore implements DataModelBackedStructure<S
   public long getRetentionTime() {
     HybridStoreConfig config = this.getHybridStoreConfig();
     if (config != null) {
-      return TopicManager.getExpectedRetentionTimeInMs(this, config);
+      return StoreUtils.getExpectedRetentionTimeInMs(this, config);
     } else {
       return DEFAULT_RT_RETENTION_TIME;
     }
